+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 
 - Delete event stream ([#203](https://github.com/commanded/eventstore/pull/203)).
+- Introduce `mix event_store.migrations` task to list migration status ((#207)[https://github.com/commanded/eventstore/pull/207]).
 
 ## v1.1.0